删除透传消息转换不相关逻辑
