+++ jetlinks-components/script-component/src/main/java/org/jetlinks/community/script/Script.java
@@ -1,3 +1,5 @@
+
+
 package org.jetlinks.community.script;
 
 import lombok.*;
@@ -13,33 +15,12 @@ public class Script {
 
     private final Object source;
 
-    private final ExecuteLimit limit;
-
-    private final ScriptLogger logger;
-
-    public boolean limitIsEnabled() {
-        return limit != null && limit.isEnabled();
-    }
-
     public static Script of(String name, String content) {
-        return Script.of(name, content, null, ExecuteLimit.defaultLimit(),null);
+        return Script.of(name, content, null);
     }
-
-    public static Script of(String name, String content, Object source) {
-        return Script.of(name, content, source, ExecuteLimit.defaultLimit(),null);
-    }
-
 
     public Script content(String content) {
-        return of(name, content, source, limit,logger);
-    }
-
-    public Script limit(ExecuteLimit limit) {
-        return of(name, content, source, limit,logger );
-    }
-
-    public Script logger(ScriptLogger logger) {
-        return of(name, content, source, limit,logger );
+        return of(name, content, source);
     }
 
 }
