+++ jetlinks-components/script-component/src/main/java/org/jetlinks/community/script/ScriptFactory.java
@@ -1,7 +1,5 @@
 package org.jetlinks.community.script;
 
-import org.jetlinks.community.script.context.ExecutionContext;
-
 import java.util.Collection;
 import java.util.Map;
 
@@ -89,12 +87,6 @@ public interface ScriptFactory {
      * @return 接口代理实现
      */
     <T> T bind(Script script,
-               Class<T> interfaceType,
-               ExecutionContext context);
-
-    default <T> T bind(Script script,
-                       Class<T> interfaceType) {
-        return bind(script, interfaceType, ExecutionContext.create());
-    }
+               Class<T> interfaceType);
 
 }
